+++ src/main/scala/au/csiro/data61/magda/search/elasticsearch/ElasticSearchImplicits.scala
@@ -24,7 +24,7 @@ object ElasticSearchImplicits {
     case (st: MultiBucketsAggregation) => st.getBuckets.asScala.map(bucket =>
       new FacetOption(
         value = bucket.getKeyAsString,
-        hitCount = Some(bucket.getDocCount.toInt)
+        hitCount = Some(bucket.getDocCount)
       )
     )
     case (_) => Seq()
